fix: resolve warnings and update swiftlint.yml
