+++ Core/Core/AuthConstants.swift
@@ -1,5 +1,5 @@
 //
-//  Constants.swift
+//  AuthConstants.swift
 //  Core
 //
 //  Created by Vladimir Chekyrta on 14.09.2022.
@@ -7,8 +7,7 @@
 
 import Foundation
 
-// TODO move it to config file and parse
-public struct Constants {
+public struct AuthConstants {
     public static let GrantTypePassword = "password"
     public static let GrantTypeRefreshToken = "refresh_token"
 }
