+++ Core/Core/Extensions/DateExtension.swift
@@ -97,7 +97,7 @@ public extension Date {
         specificFormatter.dateFormat = isCurrentYear ? "MMMM d" : "MMMM d, yyyy"
         return dueInString + specificFormatter.string(from: self)
     }
-
+    
     func isDateInNextWeek(date: Date, currentDate: Date) -> Bool {
         let calendar = Calendar.current
         guard let nextWeek = calendar.date(byAdding: .weekOfYear, value: 1, to: currentDate) else { return false }
@@ -121,19 +121,19 @@ public extension Date {
         }
         self = date
     }
-
-     init(milliseconds: Double) {
-         let now = Date()
-         let calendar = Calendar.current
-         var components = calendar.dateComponents([.year, .month, .day, .hour, .minute, .second], from: now)
-         components.nanosecond = Int((milliseconds.truncatingRemainder(dividingBy: 1)) * 1000000)
-         let seconds = Int(milliseconds)
-         components.second = seconds % 60
-         components.minute = (seconds / 60) % 60
-         components.hour = (seconds / 3600) % 24
-         let date = calendar.date(from: components) ?? Date()
-         self = date
-     }
+    
+    init(milliseconds: Double) {
+        let now = Date()
+        let calendar = Calendar.current
+        var components = calendar.dateComponents([.year, .month, .day, .hour, .minute, .second], from: now)
+        components.nanosecond = Int((milliseconds.truncatingRemainder(dividingBy: 1)) * 1000000)
+        let seconds = Int(milliseconds)
+        components.second = seconds % 60
+        components.minute = (seconds / 60) % 60
+        components.hour = (seconds / 3600) % 24
+        let date = calendar.date(from: components) ?? Date()
+        self = date
+    }
 }
 
 public enum DateStringStyle {
@@ -153,11 +153,11 @@ public extension Date {
     func secondsSinceMidnight() -> Double {
         let calendar = Calendar.current
         let components = calendar.dateComponents([.hour, .minute, .second], from: self)
-
+        
         guard let hours = components.hour, let minutes = components.minute, let seconds = components.second else {
             return 0.0
         }
-
+        
         let totalSeconds = Double(hours) * 3600.0 + Double(minutes) * 60.0 + Double(seconds)
         return totalSeconds
     }
@@ -193,7 +193,7 @@ public extension Date {
         }
         
         let date = dateFormatter.string(from: self)
-
+        
         switch style {
         case .courseStartsMonthDDYear:
             return CoreLocalization.Date.courseStarts + " " + date
@@ -229,17 +229,12 @@ public extension Date {
         case .shortWeekdayMonthDayYear:
             return (
                 dueIn ? CoreLocalization.Date.dueIn : ""
-            ) + getShortWeekdayMonthDayYear(dateFormatterString: date)
+            ) + date
         }
     }
     
     private func applyShortWeekdayMonthDayYear(dateFormatter: DateFormatter) {
-            dateFormatter.dateFormat = "MMMM d, yyyy"
-    }
-    
-    private func getShortWeekdayMonthDayYear(dateFormatterString: String) -> String {
-        let days = Calendar.current.dateComponents([.day], from: self, to: Date())
-        return dateFormatterString
+        dateFormatter.dateFormat = "MMMM d, yyyy"
     }
     
     func isCurrentYear() -> Bool {
@@ -253,7 +248,7 @@ public extension Date {
     func isEarlierThanOrEqualTo(date: Date) -> Bool {
         timeIntervalSince1970 <= date.timeIntervalSince1970
     }
-
+    
     func isLaterThanOrEqualTo(date: Date) -> Bool {
         timeIntervalSince1970 >= date.timeIntervalSince1970
     }
