+++ Discussion/Discussion/Domain/DiscussionInteractor.swift
@@ -82,10 +82,12 @@ public actor DiscussionInteractor: DiscussionInteractorProtocol {
         return try await repository.getCommentResponses(commentID: commentID, page: page)
     }
 
+    //swiftlint:disable todo
     // TODO: This Api should be updated with type GET, currently we are using this for deep linking on comment screen.
     public func getResponse(responseID: String) async throws -> UserComment {
         return try await repository.getResponse(responseID: responseID)
     }
+    //swiftlint:enable todo
 
     public func addCommentTo(threadID: String, rawBody: String, parentID: String? = nil) async throws -> Post {
         return try await repository.addCommentTo(threadID: threadID,
